Moving all block items to one folder
