+++ Blocks/BlockSprites/BlueTile.cs
@@ -1,13 +1,13 @@
 ﻿using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework;
-using Sprint2;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Sprint2_Attempt3.Blocks;
 
-namespace Sprint2_Attempt3.BlockSprites
+namespace Sprint2_Attempt3.Blocks.BlockSprites
 {
     public class BlueTile : IBlockSprite
     {
@@ -17,8 +17,8 @@ namespace Sprint2_Attempt3.BlockSprites
         public BlueTile(Texture2D tilesSet, Vector2 Pos, Rectangle diamondSource)
         {
             this.tilesSet = tilesSet;
-            this.destX = (int)Pos.X;
-            this.destY = (int)Pos.Y;
+            destX = (int)Pos.X;
+            destY = (int)Pos.Y;
             this.diamondSource = diamondSource;
         }
 
@@ -28,14 +28,14 @@ namespace Sprint2_Attempt3.BlockSprites
 
         public Rectangle DestRectangle()
         {
-            int width = this.diamondSource.Width;
-            int height = this.diamondSource.Height;
+            int width = diamondSource.Width;
+            int height = diamondSource.Height;
             return new Rectangle(destX, destY, width * 3, height * 3);
         }
 
         public void Draw(SpriteBatch spriteBatch)
         {
-            Rectangle scrRectangle = this.diamondSource;
+            Rectangle scrRectangle = diamondSource;
             spriteBatch.Draw(tilesSet, DestRectangle(), scrRectangle, Color.White);
         }
     }
