+++ Blocks/BlockSprites/UpChunk.cs
@@ -6,20 +6,21 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.ComponentModel;
+using Sprint2_Attempt3.Blocks;
 
-namespace Sprint2
+namespace Sprint2_Attempt3.Blocks.BlockSprites
 {
-    public class SideChunk : IBlockSprite
+    public class UpChunk : IBlockSprite
     {
-        private Rectangle sideChunkSource;
+        private Rectangle upChunkSource;
         private int destX, destY;
         private Texture2D blocks;
-        public SideChunk(Texture2D blocks, Vector2 Pos, Rectangle sideChunkSource)
+        public UpChunk(Texture2D blocks, Vector2 Pos, Rectangle upChunkSource)
         {
             this.blocks = blocks;
-            this.destX = (int)Pos.X;
-            this.destY = (int)Pos.Y;
-            this.sideChunkSource = sideChunkSource;
+            destX = (int)Pos.X;
+            destY = (int)Pos.Y;
+            this.upChunkSource = upChunkSource;
         }
 
         public void Update()
@@ -28,14 +29,14 @@ namespace Sprint2
 
         public Rectangle DestRectangle()
         {
-            int width = this.sideChunkSource.Width;
-            int height = this.sideChunkSource.Height;
+            int width = upChunkSource.Width;
+            int height = upChunkSource.Height;
             return new Rectangle(destX, destY, width * 3, height * 3);
         }
 
         public void Draw(SpriteBatch spriteBatch)
         {
-            Rectangle scrRectangle = this.sideChunkSource;
+            Rectangle scrRectangle = upChunkSource;
             spriteBatch.Draw(blocks, DestRectangle(), scrRectangle, Color.White);
         }
     }
