+++ Blocks/BlockSprites/WhiteStair.cs
@@ -6,8 +6,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.ComponentModel;
+using Sprint2_Attempt3.Blocks;
 
-namespace Sprint2
+namespace Sprint2_Attempt3.Blocks.BlockSprites
 {
     public class WhiteStair : IBlockSprite
     {
@@ -17,8 +18,8 @@ namespace Sprint2
         public WhiteStair(Texture2D tilesSet, Vector2 Pos, Rectangle upChunkSource)
         {
             this.tilesSet = tilesSet;
-            this.destX = (int)Pos.X;
-            this.destY = (int)Pos.Y;
+            destX = (int)Pos.X;
+            destY = (int)Pos.Y;
             this.upChunkSource = upChunkSource;
         }
 
@@ -28,14 +29,14 @@ namespace Sprint2
 
         public Rectangle DestRectangle()
         {
-            int width = this.upChunkSource.Width;
-            int height = this.upChunkSource.Height;
+            int width = upChunkSource.Width;
+            int height = upChunkSource.Height;
             return new Rectangle(destX, destY, width * 3, height * 3);
         }
 
         public void Draw(SpriteBatch spriteBatch)
         {
-            Rectangle scrRectangle = this.upChunkSource;
+            Rectangle scrRectangle = upChunkSource;
             spriteBatch.Draw(tilesSet, DestRectangle(), scrRectangle, Color.White);
         }
     }
