+++ Blocks/BlockSprites/plainTile.cs
@@ -6,20 +6,21 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.ComponentModel;
+using Sprint2_Attempt3.Blocks;
 
-namespace Sprint2
+namespace Sprint2_Attempt3.Blocks.BlockSprites
 {
-    public class DiamondTile : IBlockSprite
+    public class PlainTile : IBlockSprite
     {
-        private Rectangle diamondSource;
+        private Rectangle plainSource;
         private int destX, destY;
         private Texture2D blocks;
-        public DiamondTile(Texture2D blocks, Vector2 Pos, Rectangle diamondSource)
+        public PlainTile(Texture2D blocks, Vector2 Pos, Rectangle plainSource)
         {
             this.blocks = blocks;
-            this.destX = (int)Pos.X;
-            this.destY = (int)Pos.Y;
-            this.diamondSource = diamondSource;
+            destX = (int)Pos.X;
+            destY = (int)Pos.Y;
+            this.plainSource = plainSource;
         }
 
         public void Update()
@@ -28,14 +29,14 @@ namespace Sprint2
 
         public Rectangle DestRectangle()
         {
-            int width = this.diamondSource.Width;
-            int height = this.diamondSource.Height;
+            int width = plainSource.Width;
+            int height = plainSource.Height;
             return new Rectangle(destX, destY, width * 3, height * 3);
         }
 
         public void Draw(SpriteBatch spriteBatch)
         {
-            Rectangle scrRectangle = this.diamondSource;
+            Rectangle scrRectangle = plainSource;
             spriteBatch.Draw(blocks, DestRectangle(), scrRectangle, Color.White);
         }
     }
