+++ Blocks/BlockStates/IStateTiles.cs
@@ -1,11 +1,10 @@
-﻿using Sprint2_Attempt3;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Sprint2.BlockStates
+namespace Sprint2_Attempt3.Blocks.BlockStates
 {
     public interface IStateTiles : IGameObject
     {
