+++ Blocks/BlockStates/UpChunkState.cs
@@ -1,22 +1,15 @@
-﻿using Sprint2_Attempt3;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Sprint2
+﻿namespace Sprint2_Attempt3.Blocks.BlockStates
 {
-    public class PlainTileState : BlockStates.IStateTiles
+    public class UpChunkState : IStateTiles
     {
         private Block block;
         private Game1 game1;
 
-        public PlainTileState(Block block)
+        public UpChunkState(Block block)
         {
             this.block = block;
 
-            block.Sprite = BlockSpriteFactory.Instance.CreatePlainTile();
+            block.Sprite = BlockSpriteFactory.Instance.CreateUpChunk();
         }
         public void ChangeToDiamondTile()
         {
