+++ Blocks/BlockStates/WhiteStairState.cs
@@ -1,21 +1,21 @@
-﻿using Sprint2_Attempt3;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Sprint2
+namespace Sprint2_Attempt3.Blocks.BlockStates
 {
-    public class DiamondTileState : BlockStates.IStateTiles
+    public class WhiteStairState : IStateTiles
     {
         private Block block;
         private Game1 game1;
 
-        public DiamondTileState(Block block)
+        public WhiteStairState(Block block)
         {
             this.block = block;
-            block.Sprite = BlockSpriteFactory.Instance.CreateDiamondTile();
+
+            block.Sprite = BlockSpriteFactory.Instance.CreateWhiteStair();
         }
         public void ChangeToDiamondTile()
         {
@@ -33,6 +33,14 @@ namespace Sprint2
         {
             block.State = new SideChunkState(block);
         }
+        public void ChangeToWhiteStairs()
+        {
+            block.State = new WhiteStairState(block);
+        }
+        public void ChangeToWhiteBrick()
+        {
+            block.State = new WhiteBrickState(block);
+        }
         public void ChangeToStaircaseChuck()
         {
             block.State = new StaircaseTileState(block);
@@ -41,16 +49,6 @@ namespace Sprint2
         {
             block.State = new BlueTileState(block);
         }
-        public void ChangeToWhiteBrick()
-        {
-            block.State = new WhiteBrickState(block);
-        }
-
-        public void ChangeToWhiteStairs()
-        {
-            block.State = new WhiteStairState(block);
-        }
-
         public void ChangeToBlackBlock()
         {
             block.State = new BlackBlockState(block);
@@ -59,11 +57,11 @@ namespace Sprint2
         {
             block.State = new DotTileState(block);
         }
-
         public void Update()
         {
 
         }
+
 
     }
 }
