+++ Blocks/IBlock.cs
@@ -5,9 +5,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Xna.Framework;
-using Sprint2_Attempt3;
 
-namespace Sprint2
+namespace Sprint2_Attempt3.Blocks
 {
     public interface IBlock : IGameObject
     {
@@ -28,3 +27,7 @@ namespace Sprint2
 
     }
 }
+
+
+
+
