+++ CommandClasses/SwitchToNextBlock.cs
@@ -1,14 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework;
-using System.Diagnostics;
-using Sprint2;
-//using System.Numerics;
-//using System.Drawing;
+﻿using Sprint2_Attempt3.Blocks;
 
 namespace Sprint2_Attempt3.CommandClasses
 {
